Refactor light to gen1 and gen2 lights
